Fix issue #785: GCC 16 compiler miscompiles 4LL in memset operations
* The issue was revealed in tesseract when building fonts with bmf.c
* Replacing 4LL by (size_t)4 apparently fixes the problem.
* (I find compiler sensitivity to this very surprising)

+++ src/pix1.c
@@ -320,7 +320,7 @@ PIX  *pixd;
 
     if ((pixd = pixCreateNoInit(width, height, depth)) == NULL)
         return (PIX *)ERROR_PTR("pixd not made", __func__, NULL);
-    memset(pixd->data, 0, 4LL * pixd->wpl * pixd->h);
+    memset(pixd->data, 0, (size_t)4 * pixd->wpl * pixd->h);
     return pixd;
 }
 
@@ -386,7 +386,7 @@ PIX  *pixd;
 
     if ((pixd = pixCreateTemplateNoInit(pixs)) == NULL)
         return (PIX *)ERROR_PTR("pixd not made", __func__, NULL);
-    memset(pixd->data, 0, 4LL * pixd->wpl * pixd->h);
+    memset(pixd->data, 0, (size_t)4 * pixd->wpl * pixd->h);
     return pixd;
 }
